requirements.txt
google-cloud==0.34.0
google-cloud-storage==1.20.0
google-cloud-speech==1.2.0
gensim==3.8.1
pydub==0.23.1
